requirements.txt
lxml
beautifulsoup4
matplotlib
scikit-learn
pandas
nltk
spacy
numpy
joblib
Flask
Image
pytesseract
deskew
scikit-image
plotly
Kaleido
keras
tensorflow
sqlitedict
Biopython
pyenchant
opencv-python==4.5.5.62
elasticsearch
tqdm
tensorflow-hub
tensorflow-text
tensorflow-addons
requests
transformers[tf-cpu]
Pillow